requirements.txt
click>=8.1.2
grpcio>=1.57.0
grpcio-tools>=1.57.0
pydantic==2.7.1
python-dotenv>=0.20.0
questionary>=2.0.1
toml>=0.10.2
tabulate>=0.9.0
boto3==1.29.0
pytest==8.1.1
requests==2.31.0
toml==0.10.2
pydantic-settings==2.2.1
jwrt==1.3.1
pyjwt==2.8.0